Split numpy_interop.cpp into numpy_interop.cpp and numpy_type_interop.cpp.
